feat(transport,dmsg): browser-native WT dial for the std-Go wasm visor (build-tag fix)
The browser WebTransport dial (DialWebTransportJS, a syscall/js adapter) and the
transport-layer wt_browser.go that calls it were gated `tinygo && js && wasm`, so
the STD-Go js/wasm visor compiled the native wt_native.go path instead —
webtransport-go/quic-go, which can't open UDP in a browser. So a std-Go wasm
visor could never dial WT. (WS got away with the analogous gate because
coder/websocket has a real js implementation; quic-go does not.)

Retag by browser-vs-native, not tinygo-vs-native — the real distinction:
- transport/network: wt_native.go !tinygo → `!tinygo && !(js && wasm)` (native
  only); wt_browser.go `tinygo && js && wasm` → `js && wasm` (any browser) and it
  now also provides the Start + resolveWTViaAR stubs a browser needs (can't serve,
  no AR); wt_tinygo.go `tinygo` → `tinygo && !(js && wasm)` (embedded only) so it
  doesn't collide with wt_browser on tinygo-wasm.
- dmsg: wt.go !tinygo → `!tinygo && !(js && wasm)`; wt_js_tinygo.go
  `tinygo && js && wasm` → `js && wasm`. The shared wsAddr/wsTimeoutError helpers
  move to a new jsaddr_js.go (`js && wasm`) so both the WS (tinygo-only) and WT
  (all-js) browser adapters get them. The dmsg WS session dial is untouched
  (coder/websocket, ws.go) — the wasm visor's lifeline is unaffected.

Builds verified: native, std-Go js/wasm (cmd/wasm-visor). Symbol-consistent for
tinygo-wasm + tinygo-embedded (helpers moved to a file both include; no new
redeclarations). Enabling change — the wasm autoconnect dialing WT is the next PR.
