+++ pkg/transport/network/wt_native.go
@@ -1,4 +1,4 @@
-//go:build !tinygo
+//go:build !tinygo && !(js && wasm)
 
 // Package network pkg/transport/network/wt_native.go
 //
